requirements.txt
Django>=3.0,<4.0
pysqlite3
djangorestframework==3.13.1
django-elasticsearch-dsl==7.2.2
django-crispy-forms==1.14.0
requests==2.27.1 
djangorestframework-simplejwt==5.2.0 